moved everything into folders
vs code won't keep the association with Dockerfiles,
and the syntax highlighting it brings unless
the file is explicitly named Dockerfile.

but I can't have multiple files named Dockerfile
because that's not allowed in the first place,
and even if it were I wouldn't be able to tell
what it's actually building.